🐛 fix: improved deletion logic
Before, if the script found another template inside of the one specified for deletion, it'd immediately think that's where it ends, as it considered the first }} it found, leaving behind raw text on the page, which was visible to anyone who accessed it.

This is now fixed by counting every {{ opened and comparing with their respective closing }}, where, the moment 'aberto' (open) equals 0, that's where the template specified truly ends, allowing to proceed with its fully deletion.